Move constants and function from custom_storage.py
Move everything except CustomPartitioningSpoke from custom_storage.py
to custom_storage_helpers.py.
